GhostPost-React/pyproject.toml
[tool.poetry]
name = "ghostpost"
version = "0.1.0"
description = ""
authors = ["Derek <derekb0147@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.3"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
